+++ NEWS.md
@@ -1,6 +1,7 @@
 iBreakDown 1.0
 ----------------------------------------------------------------
 * Argument `vnames` added to `plot.break_down` to change freely labels 
+* change `cummulative` to `cumulative` [#72](https://github.com/ModelOriented/iBreakDown/issues/72)
 
 iBreakDown 0.9.9
 ----------------------------------------------------------------